🔧 Modify : refactor services structure
- expose service interface for testing
- expose model for entity relations
