+++ src/main/java/com/example/webtemplate/profile/UserProfileService.java
@@ -2,22 +2,10 @@ package com.example.webtemplate.profile;
 
 import java.util.List;
 
-import org.springframework.stereotype.Service;
+public interface UserProfileService {
 
-@Service
-class UserProfileService {
+    List<Long> getAllUserProfileIds();
 
-    private final UserProfileRepository repository;
-
-    UserProfileService(UserProfileRepository repository) {
-        this.repository = repository;
-    }
-
-    public List<Long> getAllUserProfileIds() {
-        return repository.findAll().stream()
-                .map(profile -> profile.getId())
-                .filter(id -> id != null)
-                .toList();
-    }
+    long count();
 
 }
